+++ app/src/plugins/linux/mod.rs
@@ -1,2 +1,3 @@
 pub mod users;
 pub mod tcp_listen;
+pub mod root;