+++ app/src/plugins/linux/tcp_listen/mod.rs
@@ -20,7 +20,7 @@ impl Plugin for LinuxTCPListen {
 
     fn run(&self, _config: &Config, _binary_directory: &str) -> Result<Value, String> {
         let command = "lsof -nP -iTCP -sTCP:LISTEN | grep -v COMMAND | tr -s ' ' |  cut -d ' ' -f 1,3,9 | sort";
-        match self.linux_command(&command) {
+        match self.execute_command(&command) {
             Ok(output) => self.process(&output),
             Err(e) => Err(e),
         }