+++ app/src/plugins/windows/autoruns/mod.rs
@@ -71,7 +71,7 @@ impl Plugin for WindowsAutoruns {
         // http://brianvanderplaats.com/2015/10/08/generating-json-from-csv-using-powershell/
         command.push_str(" -c -h -s -nobanner 2> $null | ConvertFrom-Csv | Select-Object Category,Enabled,Entry,\"Image Path\",\"Launch String\",MD5,SHA-1,Signer,Time | Sort-Object -Property Category,Entry | ConvertTo-Json");
 
-        match self.windows_powershell_command(&command) {
+        match self.execute_command(&command) {
             Ok(output) => self.process(&output),
             Err(e) => Err(e),
         }